Split Konsist test into multiple files
